Update README to refer to NAT-PMP's successor as something that may be supported someday
Right now the README alludes to NAT-PMP as something that'd be nice to have someday.
For informative purposes only, that specification was published as RFC 6886 so it could be used as a reference for RFC 6887: Port Control Protocol.
PCP is the modern and standardized way to politely ask a gateway to open ports and remove barriers to communication.
This role is similar to UPnP IGD, but the Port Control Protocol is designed to have more wide application and be easier to implement.
It is also useful in non-NAT scenarios to open ports that'd otherwise be blocked by a firewall.

From a semantic point of view, the Port Control Protocol should be preferred when it's available as being specifically designed for the use cases libnice is for.
It is the successor to NAT-PMP that's somewhat compatible with it (see RFC 6887 Appendix A for simple tweaks that permit clients to interoperate with both).
It is therefore mentioned as well as something that's not handled by libnice right now, but which would be nice to have.

Having this information in the README is also helpful for troubleshooting: the Port Control Protocol is supposed to alleviate the need for trickery such as STUN and TURN.
An administrator will likely find it helpful to know that libnice doesn't support NAT-PMP or PCP right now and thus troubleshooting needs to be done elsewhere.
This note also provides a hint for what low-hanging fruit a new contributor in this problem space may wish to take up.